# feat: implement browser pool optimization for JavaScript rendering