… internal/client/browser_pool_test.go
… internal/client/js_client_test.go
… internal/detector/detector_test.go
… internal/detector/spa_detector.go